scatter plot component added
